backend/requirements.txt
Flask==2.1.2
Werkzeug==2.0.3
python-dotenv
requests
yfinance